enhance(apps/frontend-manage): update tag deletion modal
